refactor: useScheduleViewModelを削除し、Modelに統合
